+++ packages/slider/index.ts
@@ -1,0 +1,2 @@
+export * from './slider.js';
+export * from './slider-thumb.js';