+++ packages/slider/oddbird-css-anchor-positioning.ts
@@ -1,0 +1,4 @@
+// This file is only here to give the Storybook Vite dev server a target
+import polyfill from '@oddbird/css-anchor-positioning/fn';
+
+export default polyfill;